Refactor SqlDialect implementation: rename `SqlDialect` to `BaseSqlDialect`, update class names for consistency (e.g., `MysqlSqlDialect` to `MysqlDialect`), remove `SqlBind`, replace with `ParameterBinder`, and rename `Route` to `DatabaseRouter`; adjust related tests and documentation accordingly.
