+++ src/PdoSqlite.php
@@ -3,7 +3,7 @@
 namespace ByJG\AnyDataset\Db;
 
 use ByJG\AnyDataset\Db\Exception\DbDriverNotConnected;
-use ByJG\AnyDataset\Db\SqlDialect\SqliteSqlDialect;
+use ByJG\AnyDataset\Db\SqlDialect\SqliteDialect;
 use ByJG\Util\Uri;
 use Override;
 
@@ -19,7 +19,7 @@ class PdoSqlite extends DbPdoDriver
     #[Override]
     public function getSqlDialectClass(): string
     {
-        return SqliteSqlDialect::class;
+        return SqliteDialect::class;
     }
 
     /**
